Rename "DefaultMetricsConsumer" to "VespaMetricsConsumer"
- This has no effect outside the config model.
